refactor(storage): move DenseChildren into children/dense.rs submodule
children.rs was growing unwieldy with both Children<T> and DenseChildren<T>
sharing the same file. Extract DenseChildren<T>, Inner<T>, DenseChildrenIter,
DenseChildrenAllIter, and all related From conversions and tests into a new
storage/src/node/children/dense.rs submodule.

children.rs becomes the module root: it declares `pub mod dense` and
re-exports the three public dense types so all existing callers are unchanged.

Also update the active design doc to reflect the new file layout, add exact
memory numbers for both ethhash and non-ethhash configurations, fix iterator
description accuracy, add `const` qualifiers to bitmap() and count() in the
API table, remove pinned line numbers in Trade-offs, and apply minor prose
fixes.
